Update app version and enhance features
The app version is updated from 1.0.0 to 1.0.1 addressing minor bug fixes. A new feature of shake detection is introduced that detects user's shaking gesture to invoke some actions. The app now creates a notification channel and requests for necessary permissions related to notifications. When the app does not have the overlay permission, it navigates the user to the settings to grant the permission. The app also schedules alarms based on certain conditions. The introduced shake detector logic is able to detect a shake gesture and trigger defined operations. If the shake gesture is detected, the app will navigate to the main activity. Finally, the shake detector will stop working after 20 minutes to save device resources. With these enhancements, the app looks forward to improving the user experience and efficiency of the app operations.

In addition, the deployment lane in the Fastfile used for automatic deployments is modified to update the version code before starting the build, contributing to better version management.