Remove shake detection feature and update version
The shake detection feature, used to launch the application, was removed as it was not aligned with app's main functionality. This involved deleting several Kotlin files handling this feature (consts.kt, extension.kt, ShakeDetectionService.kt, and ShakeDetector.kt), modifying the app's manifest XML file to remove various permissions related to this feature, and updating app's MainActivity. The app version was also updated in the build.gradle.kts file, and an inconsequential string resource was refined.